Updated server side category route

+++ server/routes/CategoryRoutes.js
@@ -3,10 +3,7 @@ const Categories = require('../model/Categories');
 const verifyToken=require("./Authentication/VerifyToken");
 
 /*Router to get all Categories from db*/
-router.get("/", verifyToken, async (req, res) => {
-    if(!req.user.userRole === 'admin')
-        return res.status(401).send("Access Denied!");
-
+router.get("/", async (req, res) => {
         await Categories.find()
         .then(Categories => res.send(Categories))
         .catch(err => res.status(400).send('Error: ' + err))
@@ -15,7 +12,6 @@ router.get("/", verifyToken, async (req, res) => {
 
 /*Router to get Category by id from db*/
 router.get("/:id", verifyToken, async (req, res) => {
-
     await Categories.findById(req.params.id)
         .then((Category) => res.send(Category))
         .catch(err => res.status(400).send("Error : " + err))
@@ -23,6 +19,9 @@ router.get("/:id", verifyToken, async (req, res) => {
 
 /*Router to add Categories to db*/
 router.post('/AddCategory',verifyToken, async (req, res) => {
+    if(!req.user.userRole === 'admin')
+    return res.status(401).send("Access Denied!");
+
     let categoryExist = await Categories.findOne({
         CategoryName: req.body.CategoryName,
       });
@@ -47,6 +46,9 @@ router.post('/AddCategory',verifyToken, async (req, res) => {
 
 /*Router to delete Categories from db*/
 router.delete("/DeleteCategory/:id",verifyToken, async (req, res) => {
+    if(!req.user.userRole === 'admin')
+        return res.status(401).send("Access Denied!");
+
     await Categories.findByIdAndDelete(req.params.id)
         .then(() => res.send("Successfully Deleted!. Category ID: ( " + req.params.id + " )"))
         .catch(err => res.status(400).send("Error : " + err))
@@ -54,6 +56,9 @@ router.delete("/DeleteCategory/:id",verifyToken, async (req, res) => {
 
 /*Router to Update Categories*/
 router.put("/UpdateCategory/:id", verifyToken, async (req, res) => {
+    if(!req.user.userRole === 'admin')
+    return res.status(401).send("Access Denied!");
+
     let categoryExist = await Categories.findOne({
         CategoryName: req.body.CategoryName,
       });
